MQTT retain and status
MQTT messages will be send as retained, so new connected subscribers
will recive last valid data and don't have to wait for new data.

New topic _status_ which will let the subscribers know if the radio is
online/connected to mqtt broker.

On initial connection to mqtt broker a last will is defined. This last
will will be published to the topic _status_ if the connection to the
mqtt broker is lost (power down, connection issues,...). The last will
publishes the message "offline".