Update main.cpp
Station name "icy-name" is not always present in datastream and, when it isn't, the previous station name appears as the prefix to the current track. This change nulls the station name prior to parsing the datastream and also suppresses the hyphen that separates the normal station name from the track name if it is absent.